requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
sentence-transformers
Pillow
torch
numpy
geopy
twilio
gunicorn
